third_party/WebKit/LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-ax-aria-attributes-expected.txt

This test checks aria-help attribute of fields in multiple fields time input UI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Non-empty value
PASS focusedElementDescription() is "Hours, AXValueDescription: 12, intValue:12, range:1-12"
PASS focusedElementDescription() is "Minutes, AXValueDescription: 34, intValue:34, range:0-59"
PASS focusedElementDescription() is "Seconds, AXValueDescription: 56, intValue:56, range:0-59"
PASS focusedElementDescription() is "Milliseconds, AXValueDescription: 789, intValue:789, range:0-999"
PASS focusedElementDescription() is "AM/PM, AXValueDescription: PM, intValue:2, range:1-2"
Empty value
PASS focusedElementDescription() is "AM/PM, AXValueDescription: --, intValue:0, range:1-2"
PASS focusedElementDescription() is "Milliseconds, AXValueDescription: ---, intValue:0, range:0-999"
PASS focusedElementDescription() is "Seconds, AXValueDescription: --, intValue:0, range:0-59"
PASS focusedElementDescription() is "Minutes, AXValueDescription: --, intValue:0, range:0-59"
PASS focusedElementDescription() is "Hours, AXValueDescription: --, intValue:0, range:1-12"

PASS successfullyParsed is true

TEST COMPLETE
